+++ GameWorld/Character.h
@@ -87,7 +87,7 @@ public:
 	 * @param health The Health of this Character
 	 * @param damage The Damage capability of this Character
 	 */
-	Character(AssetType type, const string & imageFilename, float size, const Pos2<long> & loc, string name, Reaction reaction, DoA alive, CharacterState state, Health * health, Damage * damage) ;
+	Character(AssetType type, const string & imageFilename, float size, const Pos2<float> & loc, string name, Reaction reaction, DoA alive, CharacterState state, Health * health, Damage * damage) ;
 	
 	
 	/**
@@ -96,7 +96,7 @@ public:
 	 *
 	 * @param randSeed A seed to initialize the random number generator
 	 */
-	Character(FastRand<long> rand) ;
+	Character(FastRand<int> rand) ;
 	
 	
 	/**
