+++ GameWorld/NPC.cpp
@@ -18,10 +18,10 @@ NPC::NPC(const NPC & other) :
 NPC::NPC(NPC && other) :
 	Character(std::move(other)) {}
 
-NPC::NPC(AssetType type, const string & imageFilename, float size, const Pos2<long> & loc, string name, DoA alive, CharacterState state, Health * health, Damage * damage, Reaction reaction) :
+NPC::NPC(AssetType type, const string & imageFilename, float size, const Pos2<float> & loc, string name, DoA alive, CharacterState state, Health * health, Damage * damage, Reaction reaction) :
 	Character(type, imageFilename, size, loc, name, reaction, alive, state, health, damage) {}
 
-NPC::NPC(FastRand<long> rand) :
+NPC::NPC(FastRand<int> rand) :
 	Character(rand)
 {
 	this->reaction = Reaction(rand() % 5) ;
